# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Luna REST API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/organizations/get/{id}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Получить информацию об организации",
                "description": "Возвращает подробную информацию об организации по её идентификатору",
                "operationId": "api.organizations.get",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID организации",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация об организации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ROrganization"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Организация не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/organizations/by-building/{buildingId}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Получить список организаций в здании",
                "description": "Возвращает список всех организаций, находящихся в конкретном здании",
                "operationId": "api.organizations.by_building",
                "parameters": [
                    {
                        "name": "buildingId",
                        "in": "path",
                        "description": "ID здания",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций в здании",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ROrganization"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Здание не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/organizations/by-activity/{activityId}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Получить список организаций по виду деятельности",
                "description": "Возвращает список всех организаций, которые относятся к указанному виду деятельности",
                "operationId": "api.organizations.by_activity",
                "parameters": [
                    {
                        "name": "activityId",
                        "in": "path",
                        "description": "ID вида деятельности",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций по виду деятельности",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ROrganization"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Вид деятельности не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/organizations/by-location": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Получить список организаций по местоположению",
                "description": "Возвращает список организаций, которые находятся в заданном радиусе или прямоугольной области относительно указанной точки на карте",
                "operationId": "e0b5f7052664d5b0d6b1ef87d74f1804",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "description": "Широта центральной точки",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 55.7558
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "description": "Долгота центральной точки",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 37.6176
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Тип поиска. Круг или квадрат",
                        "required": true,
                        "schema": {
                            "enum": [
                                "circle",
                                "square"
                            ],
                            "example": "circle"
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "description": "Радиус поиска в метрах (используется, если type == circle)",
                        "schema": {
                            "type": "number",
                            "format": "integer",
                            "example": 1000
                        }
                    },
                    {
                        "name": "width",
                        "in": "query",
                        "description": "Ширина квадрата в метрах (используется, если type == square)",
                        "schema": {
                            "type": "number",
                            "format": "integer",
                            "example": 1000
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список организаций по местоположению",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ROrganization"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/organizations/search/by-activity/{activityId}": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Поиск организаций по виду деятельности, включая все подкатегории",
                "description": "Ищет организации по виду деятельности, включая все подкатегории",
                "operationId": "api.organizations.search.by_activity",
                "parameters": [
                    {
                        "name": "activityId",
                        "in": "path",
                        "description": "ID вида деятельности для поиска",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Результаты поиска организаций по виду деятельности",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ROrganization"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Вид деятельности не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/organizations/search": {
            "get": {
                "tags": [
                    "Organizations"
                ],
                "summary": "Поиск организаций по названию",
                "description": "Ищет организации по названию",
                "operationId": "api.organizations.search",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Название организации для поиска",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Кафе"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Результаты поиска организаций по названию",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ROrganization"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "ErrorResponse": {
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": false
                    },
                    "error": {
                        "type": "string",
                        "example": "Описание ошибки"
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "RBuilding": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "address": {
                        "type": "string",
                        "example": "ул. Тверская, 1"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 55.7558
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": 37.6176
                    }
                },
                "type": "object"
            },
            "ROrganizationPhone": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "phone": {
                        "type": "string",
                        "example": "+7 (495) 123-45-67"
                    }
                },
                "type": "object"
            },
            "RActivity": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Еда"
                    },
                    "level": {
                        "type": "integer",
                        "example": 1
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ROrganization": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Кафе 'У Парка'"
                    },
                    "building_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00Z"
                    },
                    "building": {
                        "$ref": "#/components/schemas/RBuilding"
                    },
                    "phones": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ROrganizationPhone"
                        }
                    },
                    "activities": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RActivity"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "name": "x-api-token",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}